+++ financroo/clients.go
@@ -15,7 +15,72 @@ import (
 	obc "github.com/cloudentity/openbanking-sandbox/client"
 	httptransport "github.com/go-openapi/runtime/client"
 	"github.com/pkg/errors"
+
+	acpclient "github.com/cloudentity/acp-client-go"
 )
+
+func NewAcpClient(c Config, cfg BankConfig) (acpclient.Client, error) {
+	var (
+		tokenURL, issuerURL, authorizeURL, userinfoURL, redirectURL *url.URL
+		client                                                      acpclient.Client
+		err                                                         error
+	)
+
+	if tokenURL, err = url.Parse(fmt.Sprintf("%s/%s/%s/oauth2/token", c.ACPInternalURL, cfg.AcpClient.TenantID, cfg.AcpClient.ServerID)); err != nil {
+		return client, err
+	}
+
+	if issuerURL, err = url.Parse(fmt.Sprintf("%s/%s/%s", c.ACPInternalURL, cfg.AcpClient.TenantID, cfg.AcpClient.ServerID)); err != nil {
+		return client, err
+	}
+
+	if authorizeURL, err = url.Parse(fmt.Sprintf("%s/%s/%s/oauth2/authorize", c.ACPURL, cfg.AcpClient.TenantID, cfg.AcpClient.ServerID)); err != nil {
+		return client, err
+	}
+
+	if userinfoURL, err = url.Parse(fmt.Sprintf("%s/%s/%s/userinfo", c.ACPInternalURL, cfg.AcpClient.TenantID, cfg.AcpClient.ServerID)); err != nil {
+		return client, err
+	}
+
+	if redirectURL, err = url.Parse(fmt.Sprintf("%s/api/callback", c.UIURL)); err != nil {
+		return client, err
+	}
+
+	config := acpclient.Config{
+		ClientID:                    cfg.AcpClient.ClientID,
+		IssuerURL:                   issuerURL,
+		TokenURL:                    tokenURL,
+		AuthorizeURL:                authorizeURL,
+		UserinfoURL:                 userinfoURL,
+		RedirectURL:                 redirectURL,
+		RequestObjectSigningKeyFile: cfg.AcpClient.KeyFile,
+		Scopes:                      []string{"accounts", "openid", "offline_access"},
+		Timeout:                     cfg.AcpClient.Timeout,
+		CertFile:                    cfg.AcpClient.CertFile,
+		KeyFile:                     cfg.AcpClient.KeyFile,
+		RootCA:                      cfg.AcpClient.RootCA,
+	}
+
+	if client, err = acpclient.New(config); err != nil {
+		return client, err
+	}
+
+	return client, nil
+}
+
+type LoginClientConfig struct {
+	RootCA      string
+	Timeout     time.Duration
+	UserinfoURL string
+}
+
+func (c *Config) ToLoginClientConfig() LoginClientConfig {
+	return LoginClientConfig{
+		RootCA:      c.Login.RootCA,
+		Timeout:     c.Login.Timeout,
+		UserinfoURL: fmt.Sprintf("%s/%s/%s/userinfo", c.ACPInternalURL, c.Login.TenantID, c.Login.ServerID),
+	}
+}
 
 type LoginClient struct {
 	Config     LoginClientConfig
